refactor: introduce orchestration package and move CLI setup and reporting logic
