+++ importlinter.ini
@@ -10,6 +10,7 @@ name = Runner→Core→Modules layering
 type = layers
 layers =
     src.frame_compare.runner
+    src.frame_compare.orchestration
     src.frame_compare.core
     src.frame_compare.alignment_preview : src.frame_compare.alignment_runner : src.frame_compare.analysis : src.frame_compare.analyze_target : src.frame_compare.cache : src.frame_compare.cli_layout : src.frame_compare.cli_runtime : src.frame_compare.config_helpers : src.frame_compare.config_template : src.frame_compare.config_writer : src.frame_compare.doctor : src.frame_compare.layout_utils : src.frame_compare.media : src.frame_compare.metadata : src.frame_compare.net : src.frame_compare.planner : src.frame_compare.preflight : src.frame_compare.presets : src.frame_compare.render : src.frame_compare.report : src.frame_compare.runtime_utils : src.frame_compare.selection : src.frame_compare.slowpics : src.frame_compare.subproc : src.frame_compare.vspreview : src.frame_compare.wizard : src.frame_compare.vs : src.frame_compare.vs.env : src.frame_compare.vs.source : src.frame_compare.vs.props : src.frame_compare.vs.color : src.frame_compare.vs.tonemap
 
@@ -27,10 +28,11 @@ source_modules =
     src.frame_compare.layout_utils
     src.frame_compare.media
     src.frame_compare.metadata
+
     src.frame_compare.planner
     src.frame_compare.preflight
     src.frame_compare.presets
-    src.frame_compare.runner
+
     src.frame_compare.runtime_utils
     src.frame_compare.selection
     src.frame_compare.vspreview
@@ -56,10 +58,11 @@ source_modules =
     src.frame_compare.layout_utils
     src.frame_compare.media
     src.frame_compare.metadata
+
     src.frame_compare.planner
     src.frame_compare.preflight
     src.frame_compare.presets
-    src.frame_compare.runner
+
     src.frame_compare.runtime_utils
     src.frame_compare.selection
     src.frame_compare.vspreview
@@ -69,6 +72,5 @@ source_modules =
     src.frame_compare.vs
 forbidden_modules =
     src.frame_compare.core
-ignore_imports =
-    src.frame_compare.runner -> src.frame_compare.core
+
 as_packages = False
