+++ tests/runner/test_runner_reporting.py
@@ -27,7 +27,9 @@ def test_runner_reports_dovi_auto_enabled_simple(tmp_path: Path, monkeypatch: py
     _patch_core_helper(monkeypatch, "prepare_preflight", lambda **_kwargs: preflight)
 
     # Mock media discovery
-    monkeypatch.setattr(runner_module.media_utils, "discover_media", lambda _root: [media_root / "Alpha.mkv", media_root / "Beta.mkv"])
+    import src.frame_compare.media as media_utils
+    # Mock media discovery
+    monkeypatch.setattr(media_utils, "discover_media", lambda _root: [media_root / "Alpha.mkv", media_root / "Beta.mkv"])
 
     # Mock metadata result with RPU
     plans = [
