+++ tests/test_dovi_debug_logging.py
@@ -5,7 +5,7 @@ import json
 import pytest
 from _pytest.capture import CaptureFixture
 
-from src.frame_compare import runner
+from src.frame_compare.orchestration import setup
 from src.frame_compare.vs import tonemap as vs_tonemap
 
 
@@ -15,14 +15,14 @@ def _read_err(capsys: CaptureFixture[str]) -> str:
 
 
 def test_runner_emit_dovi_debug_skips_when_flag_false(monkeypatch: pytest.MonkeyPatch, capsys: CaptureFixture[str]) -> None:
-    monkeypatch.setenv(runner.DOVI_DEBUG_ENV_FLAG, "0")
-    runner._emit_dovi_debug({"phase": "unit"})
+    monkeypatch.setenv(setup.DOVI_DEBUG_ENV_FLAG, "0")
+    setup.emit_dovi_debug({"phase": "unit"})
     assert "[DOVI_DEBUG]" not in _read_err(capsys)
 
 
 def test_runner_emit_dovi_debug_emits_when_flag_true(monkeypatch: pytest.MonkeyPatch, capsys: CaptureFixture[str]) -> None:
-    monkeypatch.setenv(runner.DOVI_DEBUG_ENV_FLAG, "1")
-    runner._emit_dovi_debug({"phase": "unit", "value": 3})
+    monkeypatch.setenv(setup.DOVI_DEBUG_ENV_FLAG, "1")
+    setup.emit_dovi_debug({"phase": "unit", "value": 3})
     err = _read_err(capsys)
     assert "[DOVI_DEBUG]" in err
     assert json.dumps({"phase": "unit", "value": 3}) in err
